TransmitterCode
This code reads messages from the Serial Monitor and places the variables into a struct data packet. The variables like int number are unique to each key or slider and have varying corresponding values. The value of “color” keys are either 127 (when pressed down) or 0 (when lifted up), but for “brightness” and “speed” sliders, values can range anywhere between 0 to 127. Each message is assigned an ID before being sent so that prior effects are only updated when a different signal is received. The packet is sent wirelessly from the ESP32 transmitter to the ESP8266 receivers.