MIDI signal handler and the MidiMessageHandler Class in Java Processing IDE
The MidiMessageHandler class is used to determine if a key or slider on the keyboard has changed by updating and storing the most recent key ID number and value. Then, the signal handler links the MIDI keyboard to the laptop and tracks whenever a slider is moved or a key is pressed. The most recent ID number and value of a slider or key is displayed on the laptop screen and sent to the Serial Monitor to be read in Arduino IDE.